+++ src/Controller/UserController.php
@@ -34,50 +34,24 @@ class UserController extends BaseController {
      * @return \Slim\Http\Response
      */
     public function saveRegisterAction($request, $response, $args) {
-        $error = FALSE;
         $recaptcha = new \ReCaptcha\ReCaptcha($this->settings['recaptcha']['secret']);
         $resp = $recaptcha->setExpectedHostname($_SERVER['SERVER_NAME'])
             ->verify($request->getParam('g-recaptcha-response'), GeneralUtility::getUserIP());
         
         if ($resp->isSuccess() || isset($_ENV['docker'])) {
-            $userSearch = $this->em->getRepository('App\Entity\User')->findOneBy(['name' => $request->getParam('user_name'), 'hidden' => 0]);
-            $userName = $request->getParam('user_name');
-            $userPass = $request->getParam('user_pass');
-            $userPassRepeat = $request->getParam('user_pass_repeat');
-            
-            if ($userSearch instanceof User) {
-                $this->flash->addMessage('message', LanguageUtility::trans('register-flash-m1') . ';' . self::STYLE_DANGER);
-                $error = TRUE;
-            }
-                
-            if (strlen($userName) < 4) {
-                $this->flash->addMessage('message', LanguageUtility::trans('register-flash-m2') . ';' . self::STYLE_DANGER);
-                $error = TRUE;
-            }
-
-            if (strlen($userPass) < 6 || strlen($userPassRepeat) < 6) {
-                $this->flash->addMessage('message', LanguageUtility::trans('register-flash-m3') . ';' . self::STYLE_DANGER);
-                $error = TRUE;
-            }
-
-            if ($userPass !== $userPassRepeat) {
-                $this->flash->addMessage('message', LanguageUtility::trans('register-flash-m4') . ';' . self::STYLE_DANGER);
-                $error = TRUE;
-            } 
-
-            if (!$error) {
+            // if validation passed
+            if (GeneralUtility::validateUser($request)) {
                 $this->flash->addMessage('message', LanguageUtility::trans('register-flash-m5') . ';' . self::STYLE_SUCCESS);
 
                 $user = new User();
-                $user->setName($userName)
+                $user->setName($request->getParam('user_name'))
                     ->setRole($this->em->getRepository('App\Entity\Role')->findOneBy(['name' => 'member']))
-                    ->setPass($userPass);
+                    ->setPass($request->getParam('user_pass'));
                 $this->em->persist($user);
                 $this->em->flush();
                 
                 return $response->withRedirect($this->router->pathFor('user-login-' . LanguageUtility::getGenericLocale()));
             }
-            
         } else {
             $this->flash->addMessage('message', LanguageUtility::trans('register-flash-m6') . ';' . self::STYLE_DANGER);
         }
@@ -120,12 +94,12 @@ class UserController extends BaseController {
             // if user exists
             if ($userSearch instanceof User) {
                 $this->flash->addMessage('message', LanguageUtility::trans('user-save-m1') . ';' . self::STYLE_DANGER);
-            } elseif (strlen($user) < 4) {
+            } elseif (strlen($user) < $this->settings['validation']['min_user_name_length']) {
                 // if user name too short
-                $this->flash->addMessage('message', LanguageUtility::trans('user-save-m2', [4]) . ';' . self::STYLE_DANGER);
-            } elseif (strlen($pass) < 6) {
+                $this->flash->addMessage('message', LanguageUtility::trans('user-save-m2', [$this->settings['validation']['min_user_name_length']]) . ';' . self::STYLE_DANGER);
+            } elseif (strlen($pass) < $this->settings['validation']['min_password_length']) {
                 // if password too short
-                $this->flash->addMessage('message', LanguageUtility::trans('user-save-m3', [6]) . ';' . self::STYLE_DANGER);
+                $this->flash->addMessage('message', LanguageUtility::trans('user-save-m3', [$this->settings['validation']['min_password_length']]) . ';' . self::STYLE_DANGER);
             } else {
                 // if role not exists
                 if ($role === NULL) {
@@ -189,6 +163,7 @@ class UserController extends BaseController {
             // if is logged in user and allowed show_user
             $user = $this->em->getRepository('App\Entity\User')->findOneBy(['id' => $this->currentUser]);
             
+            // if user not exists or is hidden
             if ($user === NULL || $user->isHidden()) {
                 GeneralUtility::setCurrentRole('guest');
                 GeneralUtility::setCurrentUser(NULL);
@@ -222,6 +197,7 @@ class UserController extends BaseController {
         // Render view
         return $this->view->render($response, 'user/show-all.html.twig', array_merge($args, [
             'users' => $this->em->getRepository('App\Entity\User')->findAll(),
+            'roles' => $this->em->getRepository('App\Entity\Role')->findAll(),
         ]));
     }
     
@@ -256,7 +232,13 @@ class UserController extends BaseController {
         $user->setRole($role);
         $this->em->flush($user);
         
-        return $response->withRedirect($this->router->pathFor('user-show-' . LanguageUtility::getLocale(), $args));
+        $redirectPath = $this->router->pathFor('user-show-' . LanguageUtility::getLocale(), $args);
+        
+        if (is_string($request->getParam('return'))) {
+            $redirectPath = $request->getParam('return');
+        } 
+        
+        return $response->withRedirect($redirectPath);
     }
     
     /**
@@ -333,7 +315,7 @@ class UserController extends BaseController {
         // if user has 2FA enabled
         if ($user->hasTwoFactor()) {
             unset($_SESSION['pass_code']);
-            return $response->withRedirect($this->router->pathFor('user-show-' . LanguageUtility::getLocale()));
+            return $response->withRedirect($this->router->pathFor('user-show-' . LanguageUtility::getLocale(), ['name' => $user->getName()]));
         }
         
         // if empty - generate new secret and update user
@@ -408,7 +390,7 @@ class UserController extends BaseController {
         // Render view
         return $this->view->render($response, 'user/enable-two-factor.html.twig', array_merge($args, [
             'secret' => $secret,
-            'qr' => $ga->getQRCodeGoogleUrl($user->getName(), $secret, 'fs.imhh.me'),
+            'qr' => $ga->getQRCodeGoogleUrl($user->getName(), $secret, $this->settings['2fa_qrc_title']),
             'passValid' => $passValid,
             'passCode' => isset($_SESSION['pass_code']) ? $_SESSION['pass_code'] : '',
         ]));
@@ -435,7 +417,7 @@ class UserController extends BaseController {
                 GeneralUtility::setCurrentRole($user->getRole()->getName());
                 GeneralUtility::setCurrentUser($user->getId());
                 $this->logger->info("User " . $user->getId() . " logged in - UserController:twoFactor");
-                return $response->withRedirect($this->router->pathFor('user-show-' . LanguageUtility::getLocale()));
+                return $response->withRedirect($this->router->pathFor('user-show-' . LanguageUtility::getLocale(), ['name' => $user->getName()]));
             }
 
             if ($request->isPost()) {
@@ -466,7 +448,7 @@ class UserController extends BaseController {
                     GeneralUtility::setCurrentRole($user->getRole()->getName());
                     GeneralUtility::setCurrentUser($user->getId());
                     $this->logger->info("User " . $user->getId() . " logged in - UserController:twoFactor");
-                    return $response->withRedirect($this->router->pathFor('user-show-' . LanguageUtility::getLocale()));
+                    return $response->withRedirect($this->router->pathFor('user-show-' . LanguageUtility::getLocale(), ['name' => $user->getName()]));
                 }
             }
         } else {
@@ -503,7 +485,13 @@ class UserController extends BaseController {
             $this->flash->addMessage('message', LanguageUtility::trans('user-hidden-m2') . ';' . self::STYLE_DANGER);
         }
         
-        return $response->withRedirect($this->router->pathFor('user-show-' . LanguageUtility::getLocale(), $args));
+        $redirectPath = $this->router->pathFor('user-show-' . LanguageUtility::getLocale(), $args);
+        
+        if (is_string($request->getParam('return'))) {
+            $redirectPath = $request->getParam('return');
+        } 
+        
+        return $response->withRedirect($redirectPath);
     }
     
     /**
@@ -540,7 +528,14 @@ class UserController extends BaseController {
         if ($this->currentUser === $user->getId()) {
             return $response->withRedirect($this->router->pathFor('user-logout-' . LanguageUtility::getGenericLocale()));
         } else {
-            return $response->withRedirect($this->router->pathFor('user-show-' . LanguageUtility::getLocale()));
+            $user = $this->em->getRepository('App\Entity\User')->findOneBy(['id' => $this->currentUser]);
+            $redirectPath = $this->router->pathFor('user-show-' . LanguageUtility::getLocale(), ['name' => $user->getName()]);
+
+            if (is_string($request->getParam('return'))) {
+                $redirectPath = $request->getParam('return');
+            } 
+
+            return $response->withRedirect($redirectPath);
         }
     }
 }